receiver/requirements.txt
connexion[flask,uvicorn,swagger-ui]
PyYAML
pykafka
setuptools<81